infrastructure/lambda/setup/requirements.txt
cryptography>=44.0.1
pyjwt==2.10.1
requests-aws4auth==1.3.1
requests==2.32.3